requirements.txt
pandas
langchain
chainlit
transformers
torch
beautifulsoup4
selenium
python-docx
docx2txt
pinecone-client
openai
tiktoken